+++ src/collective/solr/browser/controlpanel.py
@@ -1,17 +1,22 @@
 # -*- coding: utf-8 -*-
 from plone.app.registry.browser import controlpanel
-
-try:
-    from plone.protect.interfaces import IDisableCSRFProtection
-
-    PLONE_PROTECT_INSTALLED = True
-except ImportError:
-    PLONE_PROTECT_INSTALLED = False
+from plone.protect.interfaces import IDisableCSRFProtection
 from collective.solr.interfaces import ISolrSchema, _
+from plone.restapi.controlpanels import RegistryConfigletPanel
 from Products.CMFPlone.utils import safe_unicode
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.PythonScripts.PythonScript import PythonScript
+from zope.component import adapter
 from zope.interface import alsoProvides
+from zope.interface import Interface
+
+
+@adapter(Interface, Interface)
+class SolrControlpanelAdapter(RegistryConfigletPanel):
+    schema = ISolrSchema
+    configlet_id = "SolrSettings"
+    configlet_category_id = "Products"
+    schema_prefix = "collective.solr"
 
 
 class SolrControlPanelForm(controlpanel.RegistryEditForm):
@@ -36,8 +41,7 @@ class SolrControlPanelForm(controlpanel.RegistryEditForm):
                     if not line.startswith("##")
                 ]
             )
-            if PLONE_PROTECT_INSTALLED:
-                alsoProvides(self.request, IDisableCSRFProtection)
+            alsoProvides(self.request, IDisableCSRFProtection)
         return content
 
     def applyChanges(self, data):
@@ -51,8 +55,7 @@ class SolrControlPanelForm(controlpanel.RegistryEditForm):
             portal[self.boost_script_id] = PythonScript(self.boost_script_id)
         # since we create a PythonScript in ZODB we need to
         # disable CSRF protection
-        if PLONE_PROTECT_INSTALLED:
-            alsoProvides(self.request, IDisableCSRFProtection)
+        alsoProvides(self.request, IDisableCSRFProtection)
         portal[self.boost_script_id].write(boost_script)
         return changes