+++ src/collective/solr/browser/interfaces.py
@@ -3,4 +3,4 @@ from zope.interface import Interface
 
 
 class IThemeSpecific(Interface):
-    """ marker interface that defines a zope3 browser layer """
+    """marker interface that defines a zope3 browser layer"""