+++ src/collective/solr/browser/maintenance.py
@@ -9,10 +9,6 @@ from plone.uuid.interfaces import IUUID, IUUIDAware
 from zope.interface import implementer
 from zope.component import queryUtility, queryAdapter
 
-try:
-    from collective.indexing.indexer import getOwnIndexMethod
-except ImportError:
-    getOwnIndexMethod = None
 from collective.solr.indexer import DefaultAdder
 from collective.solr.flare import PloneFlare
 from collective.solr.interfaces import ISolrConnectionManager
@@ -39,7 +35,7 @@ except ImportError:
 
 
 def timer(func=time):
-    """ set up a generator returning the elapsed time since the last call """
+    """set up a generator returning the elapsed time since the last call"""
 
     def gen(last=func()):
         while True:
@@ -51,7 +47,7 @@ def timer(func=time):
 
 
 def checkpointIterator(function, interval=100):
-    """ the iterator will call the given function for every nth invocation """
+    """the iterator will call the given function for every nth invocation"""
     counter = 0
     while True:
         counter += 1
@@ -61,10 +57,10 @@ def checkpointIterator(function, interval=100):
 
 
 def notimeout(func):
-    """ decorator to prevent long-running solr tasks from timing out """
+    """decorator to prevent long-running solr tasks from timing out"""
 
     def wrapper(*args, **kw):
-        """ wrapper with random docstring so ttw access still works """
+        """wrapper with random docstring so ttw access still works"""
         manager = queryUtility(ISolrConnectionManager)
         manager.setTimeout(None, lock=True)
         try:
@@ -77,10 +73,10 @@ def notimeout(func):
 
 @implementer(ISolrMaintenanceView)
 class SolrMaintenanceView(BrowserView):
-    """ helper view for indexing all portal content in Solr """
+    """helper view for indexing all portal content in Solr"""
 
     def mklog(self, use_std_log=False):
-        """ helper to prepend a time stamp to the output """
+        """helper to prepend a time stamp to the output"""
         write = self.request.RESPONSE.write
 
         def log(msg, timestamp=True):
@@ -93,7 +89,7 @@ class SolrMaintenanceView(BrowserView):
         return log
 
     def optimize(self):
-        """ optimize solr indexes """
+        """optimize solr indexes"""
         manager = queryUtility(ISolrConnectionManager)
         conn = manager.getConnection()
         conn.setTimeout(None)
@@ -101,7 +97,7 @@ class SolrMaintenanceView(BrowserView):
         return "solr indexes optimized."
 
     def clear(self):
-        """ clear all data from solr, i.e. delete all indexed objects """
+        """clear all data from solr, i.e. delete all indexed objects"""
         manager = queryUtility(ISolrConnectionManager)
         uniqueKey = manager.getSchema().uniqueKey
         conn = manager.getConnection()
@@ -180,11 +176,6 @@ class SolrMaintenanceView(BrowserView):
 
         for path, obj in findObjects(self.context):
             if ICheckIndexable(obj)():
-                if getOwnIndexMethod:
-                    if getOwnIndexMethod(obj, "indexObject") is not None:
-                        log("skipping indexing of %r via private method.\n" % obj)
-                        continue
-
                 count += 1
                 if count <= skip:
                     continue