+++ src/collective/solr/setuphandlers.py
@@ -1,15 +1,12 @@
 # -*- coding: utf-8 -*-
-from collective.solr.interfaces import ISolrConnectionConfig
-from collective.solr.interfaces import ISolrSchema
-from plone import api
-from plone.registry.interfaces import IRegistry
-from Products.CMFCore.utils import getToolByName
-from zope.component import getSiteManager
-from zope.component import getUtility
-from zope.component import queryUtility
-
 import logging
 
+from collective.solr.interfaces import ISolrConnectionConfig, ISolrSchema
+from plone import api
+from plone.registry import Record, field
+from plone.registry.interfaces import IRegistry
+from Products.CMFCore.utils import getToolByName
+from zope.component import getSiteManager, getUtility, queryUtility
 
 logger = logging.getLogger("collective.solr")
 PROFILE_ID = "profile-collective.solr:default"
@@ -63,3 +60,23 @@ def migrateTo4(context):
             logger.info("Added new behavior to {}".format(type_id))
 
     logger.info("Migrated to version 4")
+
+
+def migrate_to_5(context):
+    registry = getUtility(IRegistry)
+    if "collective.solr.login" not in registry.records:
+        registry_field = field.TextLine(title=u"Login")
+        registry_record = Record(registry_field)
+        registry_record.value = None
+        registry.records["collective.solr.login"] = registry_record
+    if "collective.solr.password" not in registry.records:
+        registry_field = field.TextLine(title=u"Password")
+        registry_record = Record(registry_field)
+        registry_record.value = None
+        registry.records["collective.solr.password"] = registry_record
+    if "collective.solr.use_tika" not in registry.records:
+        registry_field = field.Bool(title=u"Use Tika")
+        registry_record = Record(registry_field)
+        registry_record.value = False
+        registry.records["use_tika"] = registry_record
+    logger.info("Migrated to version 5")