+++ src/collective/solr/tests/robot/variables.py
@@ -1,3 +1,0 @@
-import pkg_resources
-
-IS_PLONE4 = pkg_resources.get_distribution("Products.CMFPlone").version.startswith("4.")