+++ src/collective/solr/tests/test_server.py
@@ -1,67 +1,40 @@
 # -*- coding: utf-8 -*-
-from Acquisition import aq_base
-from Acquisition import aq_parent
-from DateTime import DateTime
-from Missing import MV
-from Products.CMFCore.utils import getToolByName
+import unittest
+from operator import itemgetter
+from re import split
+from time import sleep
+from unittest import TestCase
 
-try:
-    from Products.CMFCore.indexing import getQueue
-except ImportError:
-    from collective.indexing.queue import getQueue
-try:
-    from Products.CMFCore.indexing import processQueue
-except ImportError:
-    from collective.indexing.queue import processQueue
-from collective.solr.dispatcher import FallBackException
-from collective.solr.dispatcher import solrSearchResults
+from Acquisition import aq_base, aq_parent
+from collective.solr.dispatcher import FallBackException, solrSearchResults
 from collective.solr.flare import PloneFlare
-from collective.solr.indexer import SolrIndexProcessor
-from collective.solr.indexer import DefaultAdder, BinaryAdder
+from collective.solr.indexer import BinaryAdder, DefaultAdder, SolrIndexProcessor
 from collective.solr.indexer import logger as logger_indexer
-from collective.solr.interfaces import ISearch
-from collective.solr.interfaces import ISolrConnectionManager
-from collective.solr.interfaces import ISolrAddHandler
+from collective.solr.interfaces import ISearch, ISolrAddHandler, ISolrConnectionManager
 from collective.solr.manager import logger as logger_manager
 from collective.solr.parser import SolrResponse
 from collective.solr.search import Search
 from collective.solr.solr import logger as logger_solr
-from collective.solr.testing import activateAndReindex
-from collective.solr.testing import HAS_LINGUAPLONE
-from collective.solr.testing import HAS_PAC
-from collective.solr.testing import LEGACY_COLLECTIVE_SOLR_FUNCTIONAL_TESTING
-from collective.solr.testing import set_attributes
+from collective.solr.testing import (
+    LEGACY_COLLECTIVE_SOLR_FUNCTIONAL_TESTING,
+    activateAndReindex,
+    set_attributes,
+)
 from collective.solr.tests.utils import numFound
-from collective.solr.utils import activate
-from collective.solr.utils import getConfig
-from operator import itemgetter
+from collective.solr.utils import activate, getConfig
+from DateTime import DateTime
+from Missing import MV
 from plone import api
-from plone.app.testing import TEST_USER_ID
-from plone.app.testing import TEST_USER_NAME
-from plone.app.testing import login
-from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID, TEST_USER_NAME, login, setRoles
 from plone.namedfile.file import NamedBlobFile
-from re import split
+from Products.CMFCore.indexing import getQueue, processQueue
+from Products.CMFCore.utils import getToolByName
 from six.moves import range
-from time import sleep
-from transaction import abort
-from transaction import commit
-from unittest import TestCase
+from transaction import abort, commit
 from zExceptions import Unauthorized
 from zope.component import getUtility, queryAdapter
 from zope.interface import implementer
 from zope.schema.interfaces import IVocabularyFactory
-
-try:
-    from Products.Archetypes.interfaces import IBaseObject
-except ImportError:
-    IBaseObject = None
-try:
-    from Products.Archetypes.config import UUID_ATTR
-except ImportError:
-    UUID_ATTR = None
-
-import unittest
 
 DEFAULT_OBJS = [
     {
@@ -107,25 +80,6 @@ DEFAULT_OBJS = [
         "depth": 0,
     },
 ]
-if not HAS_PAC:
-    DEFAULT_OBJS.extend(
-        [
-            {
-                "Title": "test_user_1_",
-                "getId": "test_user_1_",
-                "Type": "Folder",
-                "portal_type": "Folder",
-                "depth": 1,
-            },
-            {
-                "Title": "",
-                "getId": "Members",
-                "Type": "Folder",
-                "portal_type": "Folder",
-                "depth": 0,
-            },
-        ]
-    )
 
 
 @implementer(ISolrAddHandler)
@@ -174,7 +128,7 @@ class SolrMaintenanceTests(TestCase):
         return self.connection.search(q=query).read().decode("utf-8")
 
     def counts(self):
-        """ crude count of metadata records in the database """
+        """crude count of metadata records in the database"""
         info = {}
         result = self.search()
         for record in split(r'<(str|date) name="', result)[1:]:
@@ -370,14 +324,10 @@ class SolrMaintenanceTests(TestCase):
         commit()
         self.assertEqual(numFound(self.search()), 2)
         sm = self.portal.getSiteManager()
-        if api.env.plone_version() >= "5.0":
-            from plone.app.contenttypes.interfaces import IImage
+        from plone.app.contenttypes.interfaces import IImage
 
-            iface = IImage
-        else:
-            iface = IBaseObject
-        if iface:
-            sm.registerAdapter(RaisingAdder, required=(iface,), name="Image")
+        iface = IImage
+        sm.registerAdapter(RaisingAdder, required=(iface,), name="Image")
 
         self.clear_solr()
         # ignore_exceptions=False should raise the handler's exception,
@@ -483,8 +433,6 @@ class SolrErrorHandlingTests(TestCase):
         self.folder = self.portal.folder
         self.config = getConfig()
         self.port = self.config.port  # remember previous port setting
-        if HAS_LINGUAPLONE:
-            self.folder.unmarkCreationFlag()  # stop LinguaPlone from renaming
         # Prevent collective indexing queues to pile up for folder creation
         commit()
 
@@ -561,8 +509,6 @@ class SolrServerTests(TestCase):
         self.maintenance.clear()
         self.config = getConfig()
         self.search = getUtility(ISearch)
-        if HAS_LINGUAPLONE:
-            self.folder.unmarkCreationFlag()  # stop LinguaPlone from renaming
 
     def tearDown(self):
         # due to the `commit()` in the tests below the activation of the
@@ -588,9 +534,8 @@ class SolrServerTests(TestCase):
         fields.remove("geolocation")
         # remove _version_ field
         fields.remove("_version_")
-        if HAS_PAC:
-            # remove getIcon which is defined for Plone 5 only
-            fields.remove("getIcon")
+        # remove getIcon which is defined for Plone 5 only
+        fields.remove("getIcon")
 
         proc = SolrIndexProcessor(manager)
         # without explicit attributes all data should be returned
@@ -706,10 +651,10 @@ class SolrServerTests(TestCase):
         self.assertEqual(self.search("+Title:Foo").results().numFound, "1")
 
     def testSearchableTextWildcardOnNumbers(self):
-        '''
+        """
         Test if we can do wild card searches if the term consist of only numbers
         or the last character is a number.
-        '''
+        """
         conn = getUtility(ISolrConnectionManager).getConnection()
         conn.add(UID="foo", Title="foo", SearchableText="123456")
         conn.commit()
@@ -813,8 +758,6 @@ class SolrServerTests(TestCase):
         self.maintenance.reindex()
         self.config.search_pattern = None
         query = {"SearchableText": "News"}
-        if HAS_LINGUAPLONE:
-            query["Language"] = "all"
         response = solrSearchResults(**query)
         self.assertEqual(len(response), 2)
         self.assertEqual(response.response.numFound, "2")
@@ -828,8 +771,6 @@ class SolrServerTests(TestCase):
         self.config.search_pattern = u"(Title:{value}^5 OR getId:{value})"
         # for single-word searches we get both, wildcards & the custom pattern
         kw_query = {"SearchableText": "news"}
-        if HAS_LINGUAPLONE:
-            kw_query["Language"] = "all"
         response = solrSearchResults(**kw_query)
         query = response.responseHeader["params"]["q"]
         self.assertEqual(query, "(Title:(news* OR news)^5 " "OR getId:(news* OR news))")
@@ -971,8 +912,6 @@ class SolrServerTests(TestCase):
         )
 
         expected = ["/plone/events", "/plone/front-page", "/plone/news"]
-        if not HAS_PAC:
-            expected.insert(0, "/plone/Members")
         self.assertEqual(search(path={"query": "/plone", "depth": 1}), expected)
 
     def testMultiplePathSearches(self):
@@ -1049,8 +988,6 @@ class SolrServerTests(TestCase):
             "/plone/news/aggregator",
             "/plone/news/folder",
         ]
-        if not HAS_PAC:
-            expected.insert(0, "/plone/Members")
         self.assertEqual(search(["/plone/news", "/plone"], depth=1), expected)
 
     def testLogicalOperators(self):
@@ -1098,10 +1035,9 @@ class SolrServerTests(TestCase):
         wf_tool.setChainForPortalTypes(
             ("Folder", "Collection"), "simple_publication_workflow"
         )
-        if HAS_PAC:
-            wfAction(self.portal.news, "retract")
-            wfAction(self.portal.news.folder, "retract")
-            wfAction(self.portal.events, "retract")
+        wfAction(self.portal.news, "retract")
+        wfAction(self.portal.news.folder, "retract")
+        wfAction(self.portal.events, "retract")
         wfAction(self.portal.news.aggregator, "retract")
         wfAction(self.portal.events.aggregator, "retract")
         wf_tool.updateRoleMappings()
@@ -1112,9 +1048,6 @@ class SolrServerTests(TestCase):
         setRoles(self.portal, TEST_USER_ID, [])
         results = self.portal.portal_catalog(request)
         expected = ["/plone/front-page"]
-        if not HAS_PAC:
-            expected.insert(0, "/plone/Members")
-            expected.insert(1, "/plone/Members/" + TEST_USER_ID)
         self.assertEqual(sorted([r.path_string for r in results]), expected)
 
     def testEffectiveRange(self):
@@ -1734,8 +1667,12 @@ class SolrServerTests(TestCase):
         self.assertEqual(sorted([r.Title for r in results]), ["foo bar", "foo/bar"])
 
     def testForcedSimpleSearchIgnoresPoint(self):
-        self.portal.invokeFactory("Document", id="fb", title="Rundschreiben Nr. 32/2017*")
-        self.portal.invokeFactory("Document", id="fb2", title="Rundschreiben Nr 32 2017*")
+        self.portal.invokeFactory(
+            "Document", id="fb", title="Rundschreiben Nr. 32/2017*"
+        )
+        self.portal.invokeFactory(
+            "Document", id="fb2", title="Rundschreiben Nr 32 2017*"
+        )
         commit()
 
         config = getConfig()
@@ -1843,16 +1780,11 @@ class SolrServerTests(TestCase):
     def testCleanup_uid(self):
         self.maintenance.reindex()
         # low level delete to force ascync index and
-        from plone.uuid.interfaces import ATTRIBUTE_NAME
-        from plone.uuid.interfaces import IUUID
+        from plone.uuid.interfaces import ATTRIBUTE_NAME, IUUID
 
         uuid_orig = IUUID(self.portal["news"])
 
-        # Dexterity
         setattr(self.portal["news"], ATTRIBUTE_NAME, "test-solr-uid")
-        # Archetypes
-        if UUID_ATTR:
-            setattr(self.portal["news"], UUID_ATTR, "test-solr-uid")
         resp = self.search("NewsFolder")
         self.assertEqual(len(resp), 1)
         self.assertEqual(resp.results()[0]["UID"], uuid_orig)